fix: 0s and 1s conflicting with tree delimiters
tree is stored in the header which uses 0s and 1s as delimiters to specify whether the node is a leaf or not
This conflicts with the 1s inside the source file. It could be fixed using an escape sequence or returning a different datastructure from tree:to_string()
Instead, knowing that a character will only follow after a 1, we can use a flag while writing the header.

more info: https://engineering.purdue.edu/ece264/17au/hw/HW13?alt=huffman